Renamed generator for single meta class
